+++ changelog.md
@@ -141,6 +141,8 @@ quick notes
 - when adding a new or updating the quantity of an existing (will also do delete soon) menu item to the active order there is now a fading highlight rect around the updated / new item for increased visual clarity, a lil bit of added polish, and a small but useful ux improvement in easily being able to revert changes in the rare case you're unsure or forget what you'd just clicked
 - added border for window blit (explain better)
 - added text bg rect for order number in sidebar but also need to improve/change this (explain better too)
+- looots of sidebar ui changed (+ some additional extras) that ill document tomo
+    - basically is customer queue and customer state/status info